gh-1123 - Moved HLLP to new subpackage
