+++ doc/src/main/java/uk/gov/gchq/gaffer/doc/predicate/HyperLogLogPlusIsLessThanExample.java
@@ -16,7 +16,7 @@
 package uk.gov.gchq.gaffer.doc.predicate;
 
 import com.clearspring.analytics.stream.cardinality.HyperLogLogPlus;
-import uk.gov.gchq.gaffer.sketches.predicate.HyperLogLogPlusIsLessThan;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.predicate.HyperLogLogPlusIsLessThan;
 
 public class HyperLogLogPlusIsLessThanExample extends PredicateExample {
     private final HyperLogLogPlus hllp1 = new HyperLogLogPlus(15);
