+++ doc/src/main/java/uk/gov/gchq/gaffer/doc/user/walkthrough/Cardinalities.java
@@ -34,7 +34,7 @@ import uk.gov.gchq.gaffer.operation.impl.add.AddElements;
 import uk.gov.gchq.gaffer.operation.impl.generate.GenerateElements;
 import uk.gov.gchq.gaffer.operation.impl.get.GetAllElements;
 import uk.gov.gchq.gaffer.operation.impl.get.GetElements;
-import uk.gov.gchq.gaffer.sketches.predicate.HyperLogLogPlusIsLessThan;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.predicate.HyperLogLogPlusIsLessThan;
 import uk.gov.gchq.gaffer.user.User;
 import uk.gov.gchq.koryphe.impl.predicate.IsEqual;
 import java.io.IOException;
