+++ library/sketches-library/src/main/java/uk/gov/gchq/gaffer/sketches/clearspring/hyperloglogplus/predicate/HyperLogLogPlusIsLessThan.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.sketches.predicate;
+package uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.predicate;
 
 import com.clearspring.analytics.stream.cardinality.HyperLogLogPlus;
 import com.fasterxml.jackson.annotation.JsonProperty;
