+++ library/sketches-library/src/main/java/uk/gov/gchq/gaffer/sketches/clearspring/hyperloglogplus/serialisation/json/hyperloglogplus/HyperLogLogPlusJsonConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.sketches.serialisation.json.hyperloglogplus;
+package uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.json.hyperloglogplus;
 
 public final class HyperLogLogPlusJsonConstants {
     public static final String HYPER_LOG_LOG_PLUS_SERIALISER_MODULE_NAME = "HyperLogLogPlusJsonSerialiser";
