+++ library/sketches-library/src/test/java/uk/gov/gchq/gaffer/sketches/datasketches/sampling/serialisation/ReservoirNumbersUnionSerialiserTest.java
@@ -19,7 +19,7 @@ import com.yahoo.sketches.sampling.ReservoirItemsUnion;
 import org.junit.Test;
 import uk.gov.gchq.gaffer.commonutil.pair.Pair;
 import uk.gov.gchq.gaffer.serialisation.Serialiser;
-import uk.gov.gchq.gaffer.sketches.serialisation.ViaCalculatedArrayValueSerialiserTest;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.ViaCalculatedArrayValueSerialiserTest;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
