+++ library/sketches-library/src/test/java/uk/gov/gchq/gaffer/sketches/datasketches/theta/serialisation/UnionSerialiserTest.java
@@ -20,7 +20,7 @@ import com.yahoo.sketches.theta.Union;
 import org.junit.Test;
 import uk.gov.gchq.gaffer.commonutil.pair.Pair;
 import uk.gov.gchq.gaffer.serialisation.Serialiser;
-import uk.gov.gchq.gaffer.sketches.serialisation.ViaCalculatedValueSerialiserTest;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.ViaCalculatedValueSerialiserTest;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
