+++ library/spark/spark-library/src/main/java/uk/gov/gchq/gaffer/spark/serialisation/kryo/impl/HyperLogLogPlusKryoSerializer.java
@@ -22,7 +22,7 @@ import com.esotericsoftware.kryo.io.Input;
 import com.esotericsoftware.kryo.io.Output;
 import uk.gov.gchq.gaffer.core.exception.GafferRuntimeException;
 import uk.gov.gchq.gaffer.exception.SerialisationException;
-import uk.gov.gchq.gaffer.sketches.serialisation.HyperLogLogPlusSerialiser;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.HyperLogLogPlusSerialiser;
 
 /**
  * * A Kryo {@link Serializer} for a {@link HyperLogLogPlus}.
