+++ store-implementation/accumulo-store/src/main/java/uk/gov/gchq/gaffer/accumulostore/AccumuloSerialisationFactory.java
@@ -17,6 +17,7 @@
 package uk.gov.gchq.gaffer.accumulostore;
 
 import uk.gov.gchq.gaffer.serialisation.Serialiser;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.HyperLogLogPlusSerialiser;
 import uk.gov.gchq.gaffer.sketches.datasketches.frequencies.serialisation.LongsSketchSerialiser;
 import uk.gov.gchq.gaffer.sketches.datasketches.frequencies.serialisation.StringsSketchSerialiser;
 import uk.gov.gchq.gaffer.sketches.datasketches.quantiles.serialisation.DoublesUnionSerialiser;
@@ -25,7 +26,6 @@ import uk.gov.gchq.gaffer.sketches.datasketches.sampling.serialisation.Reservoir
 import uk.gov.gchq.gaffer.sketches.datasketches.sampling.serialisation.ReservoirNumbersUnionSerialiser;
 import uk.gov.gchq.gaffer.sketches.datasketches.sampling.serialisation.ReservoirStringsUnionSerialiser;
 import uk.gov.gchq.gaffer.sketches.datasketches.theta.serialisation.UnionSerialiser;
-import uk.gov.gchq.gaffer.sketches.serialisation.HyperLogLogPlusSerialiser;
 import uk.gov.gchq.gaffer.store.SerialisationFactory;
 
 /**
