+++ store-implementation/hbase-store/src/test/java/uk/gov/gchq/gaffer/hbasestore/HBaseSerialisationFactoryTest.java
@@ -20,7 +20,7 @@ import com.clearspring.analytics.stream.cardinality.HyperLogLogPlus;
 import org.junit.Test;
 import uk.gov.gchq.gaffer.exception.SerialisationException;
 import uk.gov.gchq.gaffer.serialisation.Serialiser;
-import uk.gov.gchq.gaffer.sketches.serialisation.HyperLogLogPlusSerialiser;
+import uk.gov.gchq.gaffer.sketches.clearspring.hyperloglogplus.serialisation.HyperLogLogPlusSerialiser;
 import uk.gov.gchq.gaffer.store.SerialisationFactory;
 
 import static org.junit.Assert.assertEquals;
